refactor: move prefilter backends under prefilter::backend
